+++ src/cljs/rems/actions/remark.cljs
@@ -15,7 +15,7 @@
    {:db (assoc db
                ::comment ""
                ::public false)
-    :dispatch [:rems.actions.action/set-attachment-id action-form-id nil]}))
+    :dispatch [:rems.actions.action/set-attachments action-form-id []]}))
 
 (rf/reg-sub ::comment (fn [db _] (::comment db)))
 (rf/reg-event-db ::set-comment (fn [db [_ value]] (assoc db ::comment value)))